* 'master' of 192.168.2.201:laiiihz/akuCommunity:
  调整生活缴费界面，修复生活缴费选择问题

+++ lib/model/manager/life_pay_model.dart
@@ -60,7 +60,7 @@ class DailyPaymentTypeVos {
 
 class DetailedVoList {
   int groupId;
-  int paymentPrice;
+  double paymentPrice;
   List<DetailsVoList> detailsVoList;
 
   DetailedVoList({this.groupId, this.paymentPrice, this.detailsVoList});
@@ -91,9 +91,9 @@ class DetailedVoList {
 class DetailsVoList {
   int id;
   String month;
-  int costPrice;
-  int paidPrice;
-  int totalPrice;
+  double costPrice;
+  double paidPrice;
+  double totalPrice;
   String beginDate;
   String endDate;
   String unitPriceType;
